First commit as baseline
Light UNet starting with 16 output channels. This is not the final
submission, but rather a proof-of-concept that training can converge
with a batch size of 1, provided momentum is high enough (0.9).
Description of this variation of UNet is found in comments under the
UNet class. There is weighing of boundary pixels and class balancing
involved in the binary cross entropy loss. This serves as a baseline.